// client/src/app/schema/account.ts
export class Account{
    _id: string;
    ac_id: string;
    ac_number: string;
    ac_name: string;
    official_name:string;
    balance: string;
    ac_type: string;
    status: string;
    customer_id: string;
    institution_id: string;
    balance_date: string;
    last_update_date: string;
    currency: string;
    detail: string;
    credit_limit: string;
    available_credit: string;
}